% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jokes.R
\name{jokes}
\alias{jokes}
\title{Create function to spit out joke}
\usage{
jokes(N = 1)
}
\arguments{
\item{N}{numeric, designates how many to jokes to pull}
}
\value{
A dataset with the following columns: Question, Punchline
}
\description{
jokes function outputs a joke related to the R programming langauge
}
\keyword{jokes}
